+++ 17_Spring_Jpa_Jquery_JS/src/main/java/ru/otus/spring/jquery/Spring17JqueryMain.java
@@ -1,4 +1,4 @@
-package jpa;
+package ru.otus.spring.jquery;
 
 
 import org.springframework.boot.SpringApplication;
@@ -7,11 +7,11 @@ import org.springframework.context.ApplicationContext;
 
 
 @SpringBootApplication
-public class JpaMain {
+public class Spring17JqueryMain {
 
     public static void main(String[] args) {
 
-        ApplicationContext context = SpringApplication.run(JpaMain.class);
+        ApplicationContext context = SpringApplication.run(Spring17JqueryMain.class);
 
     }
 }
