+++ 17_Spring_Jpa_Jquery_JS/src/main/java/ru/otus/spring/jquery/customExceptions/DaoException.java
@@ -1,4 +1,4 @@
-package jpa.customExceptions;
+package ru.otus.spring.jquery.customExceptions;
 
 
 public class DaoException extends RuntimeException {
