+++ 17_Spring_Jpa_Jquery_JS/src/main/java/ru/otus/spring/jquery/domain/Note.java
@@ -1,5 +1,8 @@
-package jpa.domain;
+package ru.otus.spring.jquery.domain;
 
+import lombok.*;
+
+import javax.persistence.*;
 import javax.validation.constraints.NotBlank;
 
 import static javax.persistence.FetchType.LAZY;
