+++ 17_Spring_Jpa_Jquery_JS/src/main/java/ru/otus/spring/jquery/repositories/AuthorRepository.java
@@ -1,8 +1,8 @@
-package jpa.repositories;
+package ru.otus.spring.jquery.repositories;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.otus.spring.jpa.domain.Author;
+import ru.otus.spring.jquery.domain.Author;
 
 
 public interface AuthorRepository extends JpaRepository<Author, Long> {
