+++ 17_Spring_Jpa_Jquery_JS/src/main/java/ru/otus/spring/jquery/repositories/GenreRepository.java
@@ -1,7 +1,7 @@
-package jpa.repositories;
+package ru.otus.spring.jquery.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.otus.spring.jpa.domain.Genre;
+import ru.otus.spring.jquery.domain.Genre;
 
 public interface GenreRepository extends JpaRepository<Genre, Long> {
