+++ 17_Spring_Jpa_Jquery_JS/src/main/java/ru/otus/spring/jquery/repositories/NoteRepository.java
@@ -1,7 +1,7 @@
-package jpa.repositories;
+package ru.otus.spring.jquery.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.otus.spring.jpa.domain.Note;
+import ru.otus.spring.jquery.domain.Note;
 
 import java.util.List;
