+++ 17_Spring_Jpa_Jquery_JS/src/main/java/ru/otus/spring/jquery/services/NoteService.java
@@ -1,6 +1,6 @@
-package jpa.services;
+package ru.otus.spring.jquery.services;
 
-import ru.otus.spring.jpa.domain.Note;
+import ru.otus.spring.jquery.domain.Note;
 
 import java.util.List;
