+++ 17_Spring_Jpa_Jquery_JS/src/main/java/ru/otus/spring/jquery/services/impl/BookServiceImpl.java
@@ -1,13 +1,13 @@
-package jpa.services.impl;
+package ru.otus.spring.jquery.services.impl;
 
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.otus.spring.jpa.customExceptions.DaoException;
-import ru.otus.spring.jpa.domain.Book;
-import ru.otus.spring.jpa.repositories.BookRepository;
-import ru.otus.spring.jpa.services.BookService;
+import ru.otus.spring.jquery.customExceptions.DaoException;
+import ru.otus.spring.jquery.domain.Book;
+import ru.otus.spring.jquery.repositories.BookRepository;
+import ru.otus.spring.jquery.services.BookService;
 
 import java.util.List;
