+++ 17_Spring_Jpa_Jquery_JS/src/main/java/ru/otus/spring/jquery/shell/ShellAppDelete.java
@@ -1,13 +1,13 @@
-package jpa.shell;
+package ru.otus.spring.jquery.shell;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.shell.standard.ShellComponent;
 import org.springframework.shell.standard.ShellMethod;
 import org.springframework.shell.standard.ShellOption;
-import ru.otus.spring.jpa.services.AuthorService;
-import ru.otus.spring.jpa.services.BookService;
-import ru.otus.spring.jpa.services.GenreService;
-import ru.otus.spring.jpa.services.NoteService;
+import ru.otus.spring.jquery.services.AuthorService;
+import ru.otus.spring.jquery.services.BookService;
+import ru.otus.spring.jquery.services.GenreService;
+import ru.otus.spring.jquery.services.NoteService;
 
 
 @RequiredArgsConstructor
