+++ 17_Spring_Jpa_Jquery_JS/src/main/java/ru/otus/spring/jquery/shell/ShellAppGet.java
@@ -1,17 +1,17 @@
-package jpa.shell;
+package ru.otus.spring.jquery.shell;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.shell.standard.ShellComponent;
 import org.springframework.shell.standard.ShellMethod;
 import org.springframework.shell.standard.ShellOption;
-import ru.otus.spring.jpa.domain.Author;
-import ru.otus.spring.jpa.domain.Book;
-import ru.otus.spring.jpa.domain.Genre;
-import ru.otus.spring.jpa.domain.Note;
-import ru.otus.spring.jpa.services.AuthorService;
-import ru.otus.spring.jpa.services.BookService;
-import ru.otus.spring.jpa.services.GenreService;
-import ru.otus.spring.jpa.services.NoteService;
+import ru.otus.spring.jquery.domain.Author;
+import ru.otus.spring.jquery.domain.Book;
+import ru.otus.spring.jquery.domain.Genre;
+import ru.otus.spring.jquery.domain.Note;
+import ru.otus.spring.jquery.services.AuthorService;
+import ru.otus.spring.jquery.services.BookService;
+import ru.otus.spring.jquery.services.GenreService;
+import ru.otus.spring.jquery.services.NoteService;
 
 import java.util.List;
