+++ 17_Spring_Jpa_Jquery_JS/src/main/resources/application.yml
@@ -2,7 +2,6 @@ logging:
   level:
     root: ERROR
 
-
 spring:
   main:
     allow-circular-references: true
